python/pyproject.toml
[project]
name = "live-visualization"
version = "0.0.0"
description = "Example of live visualization in Foxglove"
authors = [{ name = "Foxglove", email = "support@foxglove.dev" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["foxglove-sdk>=0.4.0", "numpy>=2.0", "requests>=2.31", "opencv-python>=4.10"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
mypy = "^1"